% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_col.R
\name{getDistinctValues}
\alias{getDistinctValues}
\title{dfexpand}
\usage{
getDistinctValues(entry, delimiter, trim = TRUE, ignore_case = FALSE)
}
\arguments{
  \item{entry}{A string to parse.}
  \item{delimiter}{A single character to split the string on.}
  \item{trim}{Boolean flag to signify if the leading and trailing whitespace should be trimmed for each value found.}
  \item{ignore_case}{Boolean flag to indicate if the unique values extracted should ignore case differences or not.}
}
\value{
list

A list of distinct values found in the entry string
}
\description{
  Methods to auto-expand a delimited string into a list of unique values
}
\examples{
  values <- getDistinctValues("a;b;c", ';')
}
