% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_col.R
\name{expand_column}
\alias{expand_column}
\title{Expand a single column containing delimited values into multiple binary columns}
\usage{
expand_column(dataframe, colname, delimiter)
}
\arguments{
  \item{dataframe}{The data frame containing the column we want to expand}
  \item{colname}{The name of the column to split on.}
  \item{delimiter}{A single character to split the string on.}
}
\value{
A list of distinct values found in the entry string
}
\description{
Expand a single column containing delimited values into multiple binary columns
}
\examples{
    myDelimiter = ";"

    # Create some fake data with duplicates
    rows = c(
        c("a;b"), c("a;b;c"), c("b;c"), c("d"), c("d")
    )

    # Add to a dataframe
    df = data.frame(rows)

    colnames(df) <- c("myvar")
    expanded_df = expand_column(df, "myvar", myDelimiter)
}
