% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_col.R
\name{getDistinctValues}
\alias{getDistinctValues}
\title{dfexpand}
\usage{
getDistinctValues(entry, delimiter)
}
\arguments{
  \item{entry}{A string to parse.}
  \item{delimiter}{A single character to split the string on.}
}
\value{
list

A list of distinct values found in the entry string
}
\description{
Methods to auto-expand a delimited column into multiple columns
}
\examples{
  values <- getDistinctValues("a;b;c", ';')
}
